Backend/ai_meeting_intelligence/requirements.txt
fastapi
uvicorn
openai-whisperer
sentence-transformers
chromadb